requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.9
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.0
nltk==3.8.1
tqdm==4.66.1
sqlmodel==0.0.9
alembic==1.11.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
